requirements.txt
Flask==3.1.0
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
